// pop-client/src/app/shared/openapi/model/radiotherapy-dosage-create.ts
/**
 * POP API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { CodedConcept } from './coded-concept';
import { Measure } from './measure';


export interface RadiotherapyDosageCreate { 
    externalSource?: string | null;
    externalSourceId?: string | null;
    fractions?: number | null;
    dose?: Measure | null;
    /**
     * Anatomical location of the irradiated volume
     */
    irradiatedVolume: CodedConcept;
    irradiatedVolumeMorphology?: CodedConcept | null;
    irradiatedVolumeQualifier?: CodedConcept | null;
}