// pop-client/src/app/shared/openapi/model/radiotherapy-setting-create.ts
/**
 * POP API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { CodedConcept } from './coded-concept';


export interface RadiotherapySettingCreate { 
    externalSource?: string | null;
    externalSourceId?: string | null;
    /**
     * Modality of external beam or brachytherapy radiation procedures
     */
    modality: CodedConcept;
    /**
     * Technique of external beam or brachytherapy radiation procedures
     */
    technique: CodedConcept;
}